lab 2 done updated 3

+++ lab2/DebugExercise/DebugExercise1.java
@@ -5,29 +5,29 @@ package DebugExercise;
  * feature work.
  */
 public class DebugExercise1 {
-    public static long divideThenRound(int top, int bottom) {
+    public static int divideThenRound(int top, int bottom) {
         if (bottom == 0) {
             throw new IllegalArgumentException();
         }
-        double quotient = (double) top / bottom;
-        long result = Math.round(quotient);
+        double quotient = (double) top / (double) bottom;
+        int result = (int) Math.round(quotient);
         return result;
     }
 
     public static void main(String[] args) {
         int t = 10;
         int b = 2;
-        long result = divideThenRound(t, b);
+        int result = divideThenRound(t, b);
         System.out.println("round(" + t + "/" + b + ")=" + result);
 
         int t2 = 9;
         int b2 = 4;
-        long result2 = divideThenRound(t2, b2);
+        int result2 = divideThenRound(t2, b2);
         System.out.println("round(" + t2 + "/" + b2 + ")=" + result2);
 
         int t3 = 3;
         int b3 = 4;
-        long result3 = divideThenRound(t3, b3);
+        int result3 = divideThenRound(t3, b3);
         System.out.println("round(" + t3 + "/" + b3 + ")=" + result3);
     }
 }
